Move useLocalStorageState to separate file
